Move sigmoid opt to math_opt
